+++ ui/src/components/uploader/QasUploader.vue
@@ -28,7 +28,7 @@
         <slot name="list" :scope="scope">
           <div class="col-12 q-col-gutter-md row">
             <div v-for="(file, index) in getFilesList(scope.files, scope)" :key="index" class="row" :class="itemClass">
-              <qas-avatar class="q-mr-sm" color="contrast-primary" icon="o_attach_file" :image="file.image" rounded :text-color="getColorFileIcon(file)" />
+              <qas-avatar class="q-mr-sm" color="contrast-primary" icon="o_attach_file" :image="file.url" rounded :text-color="getColorFileIcon(file)" />
 
               <div class="col items-center no-wrap row">
                 <div class="column no-wrap" :class="{ col: isMultiple }">
@@ -130,6 +130,10 @@ export default {
     },
 
     uploading: {
+      type: Boolean
+    },
+
+    useObjectModel: {
       type: Boolean
     }
   },
@@ -244,10 +248,15 @@ export default {
     uploaded (response) {
       const fullPath = response.xhr.responseURL.split('?').shift()
 
-      this.$emit(
-        'update:modelValue',
-        this.isMultiple ? [...this.modelValue, fullPath] : fullPath || ''
-      )
+      const objectValue = {
+        format: response.files[0].type,
+        url: fullPath,
+        name: response.files[0].name
+      }
+
+      const model = this.useObjectModel ? objectValue : fullPath
+
+      this.$emit('update:modelValue', this.isMultiple ? [...this.modelValue, model] : model || '')
 
       this.updateUploading(false)
 
@@ -286,7 +295,13 @@ export default {
       }
 
       const clonedValue = extend(true, [], this.modelValue)
-      const numberIndex = this.modelValue.findIndex(file => this.getFileName(file) === index)
+      const numberIndex = this.modelValue.findIndex(file => {
+        if (this.useObjectModel) {
+          return file.uuid === index || file.url.includes(index)
+        }
+
+        return this.getFileName(file) === index
+      })
 
       clonedValue.splice(numberIndex, 1)
 
@@ -298,14 +313,10 @@ export default {
     },
 
     getFilesList (uploadedFiles) {
-      const pathsList = Array.isArray(this.modelValue)
-        ? this.modelValue
-        : (this.modelValue ? [this.modelValue] : [])
-
       uploadedFiles = uploadedFiles.map((file, indexToDelete) => {
         return {
           isUploaded: true,
-          image: file.xhr ? file.xhr.responseURL.split('?').shift() : '',
+          url: file.xhr ? file.xhr.responseURL.split('?').shift() : '',
           name: file.name,
           progressLabel: file.__progressLabel,
           sizeLabel: file.__sizeLabel,
@@ -314,11 +325,20 @@ export default {
         }
       })
 
+      const pathsList = Array.isArray(this.modelValue)
+        ? this.modelValue
+        : (this.modelValue ? [this.modelValue] : [])
+
       const mergedList = [...pathsList, ...uploadedFiles]
 
       const files = {}
 
       mergedList.forEach(file => {
+        if (this.useObjectModel && file.uuid) {
+          files[file.uuid] = file
+          return
+        }
+
         if (file.isFailed) {
           files[file.name] = file
           return
@@ -326,12 +346,12 @@ export default {
 
         if (typeof file === 'string') {
           const fileName = this.getFileName(file)
-          files[fileName] = { image: file, isUploaded: false, name: fileName }
+          files[fileName] = { url: file, isUploaded: false, name: fileName }
           return
         }
 
-        if (file.image) {
-          const fileName = this.getFileName(file.image)
+        if (file.url) {
+          const fileName = this.getFileName(file.url)
           files[fileName] = file
         }
       })
@@ -356,6 +376,8 @@ export default {
     async addFiles () {
       const filesList = Array.from(this.hiddenInputElement.files)
       const processedFiles = []
+
+      this.$refs.hiddenInput.value = ''
 
       filesList.forEach(file => processedFiles.push(this.resizeImage(file)))
       this.uploader.addFiles(await Promise.all(processedFiles))